KT-1 ver_7 (bug fix)

+++ Src/xmllogger.h
@@ -26,7 +26,7 @@ public:
 
     void writeToLogNotFound();
 
-    void writeToLogSummary(unsigned int numberofsteps, unsigned int nodescreated, float length, double time, double cellSize);
+    void writeToLogSummary(unsigned long int numberofsteps, unsigned long int nodescreated, float length, double time, double cellSize);
 
 private:
     std::string LogFileName;
